refactor: rename project to 'vidpack'
Move code to 'src' directory, updated package name, and adjusted dependencies, scripts, and tests to reflect the new project name.
