+++ .vscode/settings.json
@@ -14,10 +14,11 @@
         "ivansaul",
         "pacman",
         "typer",
-        "vcodec"
+        "vcodec",
+        "vidpack"
     ],
     "python.testing.pytestArgs": [
-        "tests"
+        "src/tests"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
